Move hooks function to gitrepo and reduce expose repopath
